ai_vue_fastai/requirements.txt
fastapi
uvicorn
pydantic
cryptography
python-jose
python-multipart
pydantic-settings
pymysql
redis
pillow
qrcode
sqlmodel
pyspark
pymongo
motor
websockets
shortuuid